+++ TagsCloud/TagsCloud.Core/Painters/DebugPainter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 using TagsCloud.Core.Settings;
-using TagsCloud.Core.TagContainersCreators;
+using TagsCloud.Core.TagContainersProviders;
 
 namespace TagsCloud.Core.Painters;
