+++ TagsCloud/TagsCloud.Core/TagContainersProviders/TagsPreprocessors/ITagsPreprocessor.cs
@@ -1,4 +1,4 @@
-﻿namespace TagsCloud.Core.TagContainersCreators.TagsPreprocessors;
+﻿namespace TagsCloud.Core.TagContainersProviders.TagsPreprocessors;
 
 public interface ITagsPreprocessor
 {
